changes done

+++ num.py
@@ -1,7 +1,7 @@
 # Program to check if a number is even or odd
 
 # Get user input
-num = int(input("Enter a number: "))
+num = double(input("Enter a number: "))
 
 # Check if even or odd
 if num % 2 == 0:
